requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
attrs==26.1.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.135.3
filetype==1.2.0
firebase_admin==7.3.0
frozenlist==1.8.0
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.70.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.14
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
marshmallow==3.26.2
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
oauthlib==3.3.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyGithub==2.9.0
PyJWT==2.12.1
PyNaCl==1.6.2
pyparsing==3.3.2
python-dotenv==1.2.2
python-telegram-bot==22.7
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==1.0.0
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
